fix(module): remove circular dependency
