+++ projects/ngx-indexed-db/src/public_api.ts
@@ -1,2 +1,3 @@
 export * from './lib/ngxindexeddb.module';
 export * from './lib/ngx-indexed-db.service';
+export * from './lib/ngx-indexed-db.meta';
